# Server/app/internal/swaggerdocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Library Management \u0026 E-Library System — REST API for authentication, the book catalog, borrowing, reading progress, and member management.\nAll endpoints except /health and /auth/* return the shared JSON envelope: {\"success\", \"message\"|\"error\", \"data\"|\"code\"}.",
        "title": "Bibliotheca API",
        "contact": {
            "name": "Bibliotheca",
            "url": "https://github.com/dprince-03/Bibliotheca"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/books": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "List books",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.PaginatedResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "items": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/internal_modules_catalog.BookResponse"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "author_ids is required (every author must already exist); author_roles is optional and positional. available_copies is set equal to total_copies and cannot be set directly.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Create a book",
                "parameters": [
                    {
                        "description": "Book details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_catalog.CreateBookRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_catalog.BookResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "invalid body, or an author_id doesn't exist",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    },
                    "401": {
                        "description": "missing/invalid token",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    },
                    "403": {
                        "description": "librarian/admin only",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    },
                    "422": {
                        "description": "validation failed",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    }
                }
            }
        },
        "/books/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Get a book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_catalog.BookResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "invalid id",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    },
                    "404": {
                        "description": "book not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Use the assign/remove-author endpoints for authors, not this one. Changing total_copies adjusts available_copies by the same delta.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Update a book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update (all optional)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_catalog.UpdateBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_catalog.BookResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "invalid id/body",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    },
                    "401": {
                        "description": "missing/invalid token",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    },
                    "403": {
                        "description": "librarian/admin only",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    },
                    "404": {
                        "description": "book not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    },
                    "422": {
                        "description": "validation failed",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Delete a book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIResponse"
                        }
                    },
                    "400": {
                        "description": "invalid id",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    },
                    "401": {
                        "description": "missing/invalid token",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    },
                    "403": {
                        "description": "admin only",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    },
                    "404": {
                        "description": "book not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    }
                }
            }
        },
        "/books/{id}/authors": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Assign an author to a book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Author + role",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_catalog.AssignAuthorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIResponse"
                        }
                    },
                    "400": {
                        "description": "invalid id/body",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    },
                    "401": {
                        "description": "missing/invalid token",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    },
                    "403": {
                        "description": "librarian/admin only",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    },
                    "404": {
                        "description": "book or author not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    },
                    "422": {
                        "description": "validation failed",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    }
                }
            }
        },
        "/books/{id}/authors/{authorId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Refuses if it would leave the book with zero authors",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Remove an author from a book",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Author ID",
                        "name": "authorId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIResponse"
                        }
                    },
                    "400": {
                        "description": "invalid ids, or would leave zero authors",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    },
                    "401": {
                        "description": "missing/invalid token",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    },
                    "403": {
                        "description": "librarian/admin only",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    }
                }
            }
        },
        "/books/{id}/download": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Streams the file (supports range requests). 404 if the book has no uploaded file yet.",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "books"
                ],
                "summary": "Download a book's digital file",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "invalid id",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    },
                    "401": {
                        "description": "missing/invalid token",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    },
                    "404": {
                        "description": "book not found, or no file uploaded",
                        "schema": {
                            "$ref": "#/definitions/github_com_dprince-03_Bibliotheca_internal_utils.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_dprince-03_Bibliotheca_internal_utils.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_dprince-03_Bibliotheca_internal_utils.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_dprince-03_Bibliotheca_internal_utils.PaginatedResponse": {
            "type": "object",
            "properties": {
                "has_next_page": {
                    "type": "boolean"
                },
                "has_previous_page": {
                    "type": "boolean"
                },
                "items": {},
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT access token (obtained from /auth/login or /auth/register).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}